requirement.txt
click==8.1.2
Flask==2.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pika==1.2.0
python-dotenv==0.20.0
Werkzeug==2.1.1
